Move VT to same module as B, and rename module to _hart
