+++ cpys/CPS.py
@@ -1,9 +1,9 @@
-from .theta import theta_multitrack
-from .B import B_vector
-from .VT import VT
 import pandas as pd
 import numpy as np
 import xarray as xr
+
+from .theta import theta_multitrack
+from ._hart import B_vector, VT
 
 
 def compute_CPS_parameters(
